Keep run_app tests lint-clean
